src/sales_personalized_email/requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.7
python-dotenv>=1.0.0
google-generativeai>=0.3.1
langchain>=0.0.335
langchain-google-genai>=0.0.6 